+++ src/components/subscriptions/licenses/LicenseAllocationDetails.jsx
@@ -1,8 +1,6 @@
 import React from 'react';
-import LicenseAllocationNavigation from './LicenseAllocationNavigation';
-import LicenseManagementTabContentTable from './LicenseManagementTabContentTable';
 import LicenseAllocationHeader from './LicenseAllocationHeader';
-import LicenseManagementTable from './LicenseManagementTable/LicenseManagementTable';
+import LicenseManagementTable from './LicenseManagementTable';
 
 const LicenseAllocationDetails = () => (
   <>
@@ -11,15 +9,7 @@ const LicenseAllocationDetails = () => (
         <div className="mb-3">
           <LicenseAllocationHeader />
         </div>
-        <div className="row my-4">
-          <div className="col-12 col-lg-3 mb-2 mb-lg-0">
-            <LicenseAllocationNavigation />
-          </div>
-          <div className="col-12 col-lg-9">
-            <LicenseManagementTabContentTable />
-          </div>
-        </div>
-        <LicenseManagementTable/>
+        <LicenseManagementTable />
       </div>
     </div>
   </>
