+++ src/components/subscriptions/licenses/LicenseAllocationHeader.jsx
@@ -1,13 +1,9 @@
 import React, { useContext } from 'react';
-import { Row, Col } from '@edx/paragon';
-import SearchBar from '../../SearchBar';
 import { SubscriptionDetailContext } from '../SubscriptionDetailContextProvider';
-import DownloadCsvButton from '../buttons/DownloadCsvButton';
 import LicenseAllocationBanner from './LicenseAllocationBanner';
 
 const LicenseAllocationHeader = () => {
   const {
-    setSearchQuery,
     subscription,
   } = useContext(SubscriptionDetailContext);
   return (
@@ -19,19 +15,6 @@ const LicenseAllocationHeader = () => {
         {subscription.licenses?.total} licenses allocated
       </p>
       <LicenseAllocationBanner />
-      <Row className="justify-content-between">
-        <Col lg={6} xs={12} className="mb-2">
-          <SearchBar
-            placeholder="Search by email..."
-            onSearch={searchQuery => setSearchQuery(searchQuery)}
-            onClear={() => setSearchQuery(null)}
-            inputProps={{ 'data-hj-suppress': true }}
-          />
-        </Col>
-        <Col xs={12} className="col-lg-auto">
-          <DownloadCsvButton />
-        </Col>
-      </Row>
     </>
   );
 };
